requirements.txt
fastapi
uvicorn
slowapi
pydantic
gitpython
pyyaml
